--- src/main/java/duke/Duke.java
@@ -1,45 +1,0 @@
-package duke;
-
-import duke.command.Command;
-import duke.data.TaskList;
-import duke.data.exception.DukeException;
-import duke.parser.Parser;
-import duke.storage.Storage;
-import duke.ui.Ui;
-
-/**
- * Represents Duke application.
- */
-public class Duke {
-
-    private Storage storage;
-    private TaskList tasks;
-    private Ui ui;
-
-    /**
-     * Constructor for Duke.
-     *
-     * @throws DukeException If loading of saved tasks fails.
-     */
-    public Duke() throws DukeException {
-        ui = new Ui();
-        storage = new Storage("src/main/java/duke/data/duke.txt");
-        try {
-            tasks = new TaskList(storage.load());
-        } catch (DukeException e) {
-            ui.showError(e.getMessage());
-            tasks = new TaskList();
-        }
-    }
-
-    public String getResponse(String input) {
-        try {
-            Command c = Parser.parse(input);
-            return c.execute(tasks, ui, storage);
-        } catch (Exception e) {
-            return e.getMessage();
-        }
-    }
-
-}
-
