--- src/main/java/duke/command/ExitCommand.java
@@ -1,31 +1,0 @@
-package duke.command;
-
-import duke.data.TaskList;
-import duke.data.exception.DukeException;
-import duke.storage.Storage;
-import duke.ui.Ui;
-
-/**
- * Exits Duke after saving state.
- */
-public class ExitCommand extends Command {
-
-    /**
-     * Executes the command given by the user.
-     *
-     * @param tasks   to be modified
-     * @param ui      to display changes
-     * @param storage to interact with as necessary
-     * @return Response string.
-     * @throws DukeException
-     */
-    @Override
-    public String execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        storage.save(tasks.tasksToStr());
-        return ui.showGoodbye();
-    }
-    @Override
-    public boolean isExit() {
-        return true;
-    }
-}
