--- src/main/java/duke/data/exception/DukeException.java
@@ -1,10 +1,0 @@
-package duke.data.exception;
-
-/**
- * Represents a Duke application exception.
- */
-public class DukeException extends Exception {
-    public DukeException(String err) {
-        super(err);
-    }
-}
