+++ src/main/java/reborn/data/exception/RebornException.java
@@ -1,0 +1,10 @@
+package reborn.data.exception;
+
+/**
+ * Represents a Duke application exception.
+ */
+public class RebornException extends Exception {
+    public RebornException(String err) {
+        super(err);
+    }
+}
