+++ src/main/java/reborn/ui/MainWindow.java
@@ -1,0 +1,66 @@
+package reborn.ui;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.Button;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.control.TextField;
+import javafx.scene.image.Image;
+import javafx.scene.layout.AnchorPane;
+import javafx.scene.layout.VBox;
+import reborn.Reborn;
+/**
+ * Controller for MainWindow. Provides the layout for the other controls.
+ */
+public class MainWindow extends AnchorPane {
+    @FXML
+    private ScrollPane scrollPane;
+    @FXML
+    private VBox dialogContainer;
+    @FXML
+    private TextField userInput;
+    @FXML
+    private Button sendButton;
+
+    private Reborn reborn;
+
+    private Image userImage = new Image(this.getClass().getResourceAsStream("/images/Verde.png"));
+    private Image dukeImage = new Image(this.getClass().getResourceAsStream("/images/Reborn.png"));
+
+    /**
+     * Initializes MainWindow.
+     */
+    @FXML
+    public void initialize() {
+        scrollPane.vvalueProperty().bind(dialogContainer.heightProperty());
+        showWelcome();
+    }
+
+    public void setDuke(Reborn d) {
+        reborn = d;
+    }
+
+    /**
+     * Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
+     * the dialog container. Clears the user input after processing.
+     */
+    @FXML
+    private void handleUserInput() {
+        String input = userInput.getText();
+        String response = reborn.getResponse(input);
+
+        dialogContainer.getChildren().addAll(
+                DialogBox.getUserDialog(input, userImage),
+                DialogBox.getRebornDialog(response, dukeImage)
+        );
+        userInput.clear();
+    }
+
+    /**
+     * Displays welcome message.
+     */
+    public void showWelcome() {
+        dialogContainer.getChildren().add(DialogBox.getRebornDialog(
+                "Reborn here. What do you want...", dukeImage));
+    }
+
+}
