+++ text-ui-test/EXPECTED.TXT
@@ -3,7 +3,7 @@
 | | | | | | | |/ / _ \
 | |_| | |_| |   <  __/
 |____/ \__,_|_|\_\___|
-Hello! I'm duke.Duke. How may I be of assistance?
+Hello! I'm reborn.Reborn. How may I be of assistance?
 ---------------------------
 Invalid input.
 ---------------------------
